neotoma-client/requirements.txt
click
requests
python-hglib
mercurial
